bot/requirements.txt
python-telegram-bot==20.7
paho-mqtt>=2.0.0